requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
Flask-Cors==3.0.10
bcrypt==4.3.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
gunicorn==21.2.0
